매물 찜하기 기능 구현 및 관련 UI 추가

+++ api/src/main/java/com/zipsoon/api/application/estate/ScoreService.java
@@ -125,7 +125,7 @@ public class ScoreService {
 
         // 비활성화된 점수 유형 필터링
         List<ScoreDto> filteredFactors = scoreFactors.stream()
-            .filter(factor -> !disabledScoreTypeIds.contains(factor.getScoreTypeId()))
+            .filter(factor -> !disabledScoreTypeIds.contains(factor.getScoreTypeId().intValue()))
             .toList();
 
         log.debug("Filtered score factors: original={}, filtered={}, disabled={}",
